# packages/api-client/openapi/reboxe.yaml
openapi: 3.1.0
info:
  title: REBOXE API
  version: 1.4.0
servers:
  - url: http://127.0.0.1:3001
paths:
...
  /v1/notifications:
    get:
      operationId: listInAppNotifications
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: In-app notification inbox } }
  /v1/notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      security: [{ bearerAuth: [] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: string } }]
      responses: { "200": { description: Owned notification marked read } }
  /v1/legal/{slug}:
    get:
      operationId: getEffectiveLegalArtifact
      parameters: [{ in: path, name: slug, required: true, schema: { type: string } }]
      responses: { "200": { description: Immutable effective legal artifact } }
  /v1/legal/{slug}/accept:
    post:
      operationId: acceptLegalArtifact
      security: [{ bearerAuth: [] }]
      parameters: [{ in: path, name: slug, required: true, schema: { type: string } }]
      responses: { "201": { description: Immutable versioned legal acceptance } }
  /v1/support/tickets:
    post:
      operationId: createSupportTicket
      security: [{ bearerAuth: [] }]
      responses: { "201": { description: Owned support ticket } }
  /v1/support/tickets/{id}:
    get:
      operationId: getOwnedSupportTicket
      security: [{ bearerAuth: [] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: string } }]
      responses: { "200": { description: Owned support ticket } }
  /v1/admin/support/tickets:
    get:
      operationId: listAdminSupportTickets
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: AAL2 support queue } }
  /v1/admin/support/tickets/{id}/reply:
    post:
      operationId: replyAdminSupportTicket
      security: [{ bearerAuth: [] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: string } }]
      responses: { "201": { description: Audited support response and status } }
  /v1/privacy/requests:
    post:
      operationId: createPrivacyRequest
      security: [{ bearerAuth: [] }]
      responses: { "201": { description: Step-up privacy request with exception receipt } }
  /v1/privacy/requests/{id}:
    get:
      operationId: getOwnedPrivacyRequest
      security: [{ bearerAuth: [] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: string } }]
      responses: { "200": { description: Owned privacy request } }
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...